+++ api/src/user/guard/session-auth.guard.ts
@@ -1,11 +1,9 @@
 import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common';
-import { GqlExecutionContext } from '@nestjs/graphql';
 
 @Injectable()
 export class SessionAuthGuard implements CanActivate {
   canActivate(context: ExecutionContext): boolean {
-    const ctx = GqlExecutionContext.create(context);
-    const request = ctx.getContext().req;
+    const request = context.switchToHttp().getRequest();
     // attach session userId to request
     if (request.session.userId) {
       request.userId = request.session.userId;