+++ api/src/user/user.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
+import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { UserDAO } from './user.dao';
 import { AuthService } from './auth.service';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
+import { AuthController } from './auth.controller';
 import { SessionAuthGuard } from './guard/session-auth.guard';
 import { RolesGuard } from './guard/roles.guard';
-import { AuthResolver } from './auth.resolver';
-import { UserResolver } from './user.resolver';
 
 @Module({
   imports: [
@@ -19,15 +19,8 @@ import { UserResolver } from './user.resolver';
       },
     }),
   ],
-  providers: [
-    UserService,
-    UserDAO,
-    AuthService,
-    RolesGuard,
-    SessionAuthGuard,
-    AuthResolver,
-    UserResolver,
-  ],
+  controllers: [UserController, AuthController],
+  providers: [UserService, UserDAO, AuthService, RolesGuard, SessionAuthGuard],
   exports: [RolesGuard, SessionAuthGuard, UserService, AuthService],
 })
 export class UserModule {}