+++ client/pages/attacklibrary.js
@@ -103,31 +103,12 @@ const AttackLibrary = (props) => {
             setAttacks(tempList)
         } else {
             setTimeout(async () => {
-                await fetchAPI(
-                    `
-                    query searchExploitByName($pageNo: Float!, $input: SearchExploitsQueryDto!) {
-                        searchExploitByName(page: $pageNo, queryInfo: $input) {
-                          id
-                          name
-                          authorName
-                          description
-                          targetAddr
-                          targetNames
-                          verified
-                        }
-                      }                      
-                    `,
-                    {
-                        pageNo: page,
-                        input: {
-                            name: search
-                        }
-                    }
-                )
-                    .then(({ data }) => {
+                await fetchAPI(`/exploit/search/${page}`)
+                    .then((res) => res.json())
+                    .then((data) => {
                         setPage(page + 1)
                         setAttacks(() => {
-                            return [...attacks, ...data["searchExploitByName"]]
+                            return [...attacks, ...data.data]
                         })
                     })
                     .catch((error) => {